Create Design pdf viewer
There is a list of 26 character heights aligned by index to their letters. For example, 'a' is at index 0 and 'z' is at 25 index . There will also be a string. Using the letter heights given, determine the area of the rectangle highlight in mm^2  assuming all letters are 1mm wide.