Create Angry Professor
A Discrete Mathematics professor has a class of students. Frustrated with their lack of discipline, the professor decides to cancel class if fewer than some number of students are present when class starts. Arrival times go from on time (arrivalTime<=0) to arrived late (arrivalTime>0).

Given the arrival time of each student and a threshold number of attendees, determine if the class is cancelled.

EXAMPLE:
n (no. of students)=5
k (threshold number of students)=3
a=[-2,-1,0,1,2]
The first  students arrived on. The last  were late. The threshold is  students, so class will go on. Return YES